refactor: replace extend of SearchContextComponent with FC and useContext
