+++ src/app/providers/search/lib/useSearchContext.ts
@@ -1,0 +1,8 @@
+import { useContext } from 'react';
+import { SearchContext } from '../ui/SearchProvider';
+
+export const useSearchContext = () => {
+  const context = useContext(SearchContext);
+
+  return context;
+};
