+++ src/app/providers/search/ui/SearchProvider.tsx
