+++ src/widgets/header/ui/Header.tsx
@@ -1,19 +1,17 @@
-import { SearchContextComponent } from '@/app/providers/search';
+import { useSearchContext } from '@/app/providers/search';
 import { SearchForm } from '@/features/search-form';
 import styles from './Header.module.scss';
 
-export class Header extends SearchContextComponent {
-  render() {
-    const { setQuery, query, isLoading } = this.context;
+export const Header = () => {
+  const { setQuery, query, isLoading } = useSearchContext();
 
-    return (
-      <header className={styles.header}>
-        <SearchForm
-          defaultValue={query}
-          disabled={isLoading}
-          onSubmit={setQuery}
-        />
-      </header>
-    );
-  }
-}
+  return (
+    <header className={styles.header}>
+      <SearchForm
+        defaultValue={query}
+        disabled={isLoading}
+        onSubmit={setQuery}
+      />
+    </header>
+  );
+};
